Add cdn for jquery, css and view,models for post issue function but not complete
