+++ chatting/assets/js/custom.js
@@ -1,0 +1,3 @@
+jQuery(document).ready(function ($) {
+    $('#my_modal_runner').DjangoModalRunner();
+});
