+++ chatting/functional_tests/issue/base.py
@@ -1,0 +1,12 @@
+__author__ = 'judelee'
+from selenium import webdriver
+from django.contrib.staticfiles.testing import StaticLiveServerTestCase
+
+
+class FunctionalTest(StaticLiveServerTestCase):
+    def setUp(self):
+        self.browser = webdriver.Firefox()
+        self.browser.implicitly_wait(3)
+
+    def tearDown(self):
+        self.browser.quit()
