+++ chatting/functional_tests/issue/test_post_issue.py
@@ -1,0 +1,9 @@
+__author__ = 'judelee'
+
+from base import FunctionalTest
+
+
+class PostissueTest(FunctionalTest):
+
+    def test_post_issue(self):
+        pass
