+++ chatting/team/tests.py
@@ -16,5 +16,5 @@ class TeamTest(TestCase):
     def test_index_returns_correct_html(self):
         request = HttpRequest()
         response = index(request)
-        expected_html = render_to_string('base.html')
+        expected_html = render_to_string('common/base.html')
         self.assertEqual(response.content.decode(), expected_html)
