+++ chatting/templates/common/base.html
@@ -1,0 +1,52 @@
+<!doctype html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <title>issue chat</title>
+    {% load staticfiles %}
+    <link rel="stylesheet" href="{% static 'css/style.css' %}">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css">
+
+</head>
+<body>
+    <div id="wrap" class="clearfix">
+        <div id="header" class="clearfix">
+            <div class="logo">
+                <h1>TeamMario</h1>
+            </div>
+            <div class="search">
+                <span><input type="text" name="q" id="q" placeholder="search here"></span>
+                <span class="clearfix"><input type="button" id="btn_search" value="Search"></span>
+            </div>
+            <div class="my_information">
+                My Information
+            </div>
+        </div>
+        <div id="nav">
+            <div class='post_issue'>
+                 <a id="my_modal_runner" class='btn' href='{% url "post_issue_form"%}'>Post Issue</a>
+            </div>
+            <div class='sorted_issue_list'>
+                <h2>Favorite Issues</h2>
+                <ul>
+                    <li>no search result(14)</li>
+                    <li>project plan</li>
+                    <li>how to good mentoring(22)</li>
+                    <li>software maestro</li>
+                <ul>
+            </div>
+        </div>
+        <div id="container">
+	        Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>Django<br><br><br><br><br>
+        </div>
+    </div>
+
+    <script src="http://code.jquery.com/jquery.min.js"></script>
+    <script src="/static/js/custom.js"></script>
+    <!-- Latest compiled and minified JavaScript -->
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.5/js/bootstrap.min.js"></script>
+</body>
+</html>
+
