+++ chatting/templates/issue/post_issue_form.html
@@ -1,0 +1,1 @@
+{% extends 'base.html' %}
