+++ chatting/tox.ini
@@ -1,0 +1,3 @@
+[flake8]
+ignore = E501
+exclude = team/migrations/
